+++ packages/frontend/editor-ui/src/features/settings/users/views/SettingsUsersView.test.ts
@@ -160,16 +160,12 @@ describe('SettingsUsersView', () => {
 			.mockResolvedValue({ link: 'https://example.com/reset/123' });
 		usersStore.updateOtherUserSettings = vi.fn().mockResolvedValue(undefined);
 		usersStore.updateGlobalRole = vi.fn().mockResolvedValue(undefined);
-		usersStore.updateEnforceMfa = vi.fn().mockResolvedValue(undefined);
 		usersStore.generateInviteLink = vi
 			.fn()
 			.mockResolvedValue({ link: 'https://example.com/signup?token=generated-token' });
 
 		settingsStore.isSmtpSetup = true;
-		settingsStore.isMFAEnforced = false;
-		settingsStore.settings.enterprise = {
-			mfaEnforcement: true,
-		} as FrontendSettings['enterprise'];
+		settingsStore.settings.enterprise = {} as FrontendSettings['enterprise'];
 		settingsStore.settings.enterprise[EnterpriseEditionFeature.AdvancedPermissions] = true;
 		ssoStore.isSamlLoginEnabled = false;
 	});
@@ -179,42 +175,6 @@ describe('SettingsUsersView', () => {
 		// Reset mock implementation to default (feature flag disabled)
 		mockIsVariantEnabled.mockReset();
 		mockIsVariantEnabled.mockReturnValue(false);
-	});
-
-	it('should turn enforcing mfa on', async () => {
-		const { getByTestId } = renderComponent();
-
-		const actionSwitch = getByTestId('enable-force-mfa');
-		expect(actionSwitch).toBeInTheDocument();
-
-		await userEvent.click(actionSwitch);
-
-		expect(usersStore.updateEnforceMfa).toHaveBeenCalledWith(true);
-	});
-
-	it('should turn enforcing mfa off', async () => {
-		settingsStore.isMFAEnforced = true;
-		const { getByTestId } = renderComponent();
-
-		const actionSwitch = getByTestId('enable-force-mfa');
-		expect(actionSwitch).toBeInTheDocument();
-
-		await userEvent.click(actionSwitch);
-
-		expect(usersStore.updateEnforceMfa).toHaveBeenCalledWith(false);
-	});
-
-	it('should handle MFA enforcement error', async () => {
-		usersStore.updateEnforceMfa = vi.fn().mockRejectedValue(new Error('MFA update failed'));
-
-		const { getByTestId } = renderComponent();
-
-		const actionSwitch = getByTestId('enable-force-mfa');
-		await userEvent.click(actionSwitch);
-
-		await waitFor(() => {
-			expect(mockToast.showError).toHaveBeenCalledWith(expect.any(Error), expect.any(String));
-		});
 	});
 
 	it('should handle missing user settings in SSO actions', async () => {
@@ -1102,33 +1062,6 @@ describe('SettingsUsersView', () => {
 				filter: {
 					fullText: '',
 				},
-			});
-		});
-
-		it('should handle MFA enforcement with partial success', async () => {
-			let callCount = 0;
-			usersStore.updateEnforceMfa = vi.fn().mockImplementation(async () => {
-				callCount++;
-				if (callCount === 1) {
-					throw Error('First attempt failed');
-				}
-				return await Promise.resolve();
-			});
-
-			const { getByTestId } = renderComponent();
-
-			const actionSwitch = getByTestId('enable-force-mfa');
-			await userEvent.click(actionSwitch);
-
-			await waitFor(() => {
-				expect(mockToast.showError).toHaveBeenCalledWith(expect.any(Error), expect.any(String));
-			});
-
-			// Try again
-			await userEvent.click(actionSwitch);
-
-			await waitFor(() => {
-				expect(usersStore.updateEnforceMfa).toHaveBeenCalledTimes(2);
 			});
 		});
